refactor: use chakra showcase next image
didn't realise that were was a chakra next image component already built
for the showcase page. i have used that now in the home page
references #136
