+++ pages/showcase.tsx
@@ -6,7 +6,7 @@ import { SkipNavContent, SkipNavLink } from '@chakra-ui/skip-nav'
 import { useColorModeValue } from '@chakra-ui/color-mode'
 import SEO from 'components/seo'
 import Header from 'components/header'
-import ChakraNextImage from 'components/showcase/chakra-next-image'
+import ChakraNextImage from 'components/chakra-next-image'
 import DiscordStrip from 'components/discord-strip'
 import { AdBanner } from 'components/chakra-pro/ad-banner'
 import ShowcaseGridItem from 'components/showcase/showcase-grid-item'
