+++ src/components/chakra-next-image.tsx
